pyproject.toml
[tool.poetry]
name = "clean-django"
version = "1.0.0"
description = ""
authors = ["Ashot Svazyan <ashot.svazyan@yandex.ru>"]
readme = "README.md"
repository = "https://github.com/TheWitcher1991/clean-django"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2"
djangorestframework = "^3.16.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"
isort = "^6.0.0"


[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''


[tool.isort]
profile = "black"
line_length = 120
multi_line_output = 3
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
